5a_: long -> int32

+++ src/cmd/5a_/lex.c
@@ -488,7 +488,7 @@ zname(char *n, int t, int s)
 void
 zaddr(Gen *a, int s)
 {
-	long l;
+	int32 l;
 	int i;
 	char *n;
 	Ieee e;
